Split DynamicListPreset and ProgressBarDialogIndeterminate
